WFLY-12722 Upgrade Infinispan to 10.1.8.Final
