+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/DefaultCacheContainer.java
@@ -23,11 +23,8 @@
 package org.jboss.as.clustering.infinispan;
 
 import java.util.Collections;
-import java.util.EnumSet;
 import java.util.List;
 import java.util.function.Function;
-
-import javax.security.auth.Subject;
 
 import org.infinispan.Cache;
 import org.infinispan.configuration.cache.Configuration;
@@ -44,13 +41,15 @@ import org.wildfly.clustering.infinispan.spi.CacheContainer;
  * EmbeddedCacheManager decorator that overrides the default cache semantics of a cache manager.
  * @author Paul Ferraro
  */
-public class DefaultCacheContainer extends AbstractDelegatingEmbeddedCacheManager implements CacheContainer, EmbeddedCacheManagerAdmin {
+public class DefaultCacheContainer extends AbstractDelegatingEmbeddedCacheManager implements CacheContainer {
 
     private final ServiceValueRegistry<Cache<?, ?>> registry;
     private final Function<String, ServiceName> serviceNameFactory;
+    private final EmbeddedCacheManagerAdmin administrator;
 
     public DefaultCacheContainer(EmbeddedCacheManager container, ServiceValueRegistry<Cache<?, ?>> registry, Function<String, ServiceName> serviceNameFactory) {
         super(container);
+        this.administrator = new DefaultCacheContainerAdmin(this);
         this.registry = registry;
         this.serviceNameFactory = serviceNameFactory;
     }
@@ -124,43 +123,12 @@ public class DefaultCacheContainer extends AbstractDelegatingEmbeddedCacheManage
 
     @Override
     public EmbeddedCacheManagerAdmin administration() {
-        return this;
-    }
-
-    @Override
-    public EmbeddedCacheManagerAdmin withFlags(AdminFlag... flags) {
-        return this;
-    }
-
-    @Override
-    public EmbeddedCacheManagerAdmin withFlags(EnumSet<AdminFlag> flags) {
-        return this;
-    }
-
-    @Override
-    public EmbeddedCacheManagerAdmin withSubject(Subject subject) {
-        return this;
-    }
-
-    @Override
-    public <K, V> Cache<K, V> createCache(String name, String template) {
-        return this.createCache(name, this.getCacheConfiguration(name));
-    }
-
-    @Override
-    public <K, V> Cache<K, V> getOrCreateCache(String name, String template) {
-        return this.getOrCreateCache(name, this.getCacheConfiguration(name));
+        return this.administrator;
     }
 
     @Override
     public synchronized <K, V> Cache<K, V> createCache(String name, Configuration configuration) {
-        this.defineConfiguration(name, configuration);
-        return this.getCache(name);
-    }
-
-    @Override
-    public synchronized <K, V> Cache<K, V> getOrCreateCache(String name, Configuration configuration) {
-        return (this.getCacheConfiguration(name) != null) ? this.createCache(name, configuration) : this.getCache(name);
+        return this.administrator.createCache(name, configuration);
     }
 
     @Override
