+++ pom.xml
@@ -364,7 +364,7 @@
         <version.org.hibernate.search>5.10.7.Final</version.org.hibernate.search>
         <version.org.hibernate.validator>6.0.20.Final</version.org.hibernate.validator>
         <version.org.hornetq>2.4.7.Final</version.org.hornetq>
-        <version.org.infinispan>10.1.6.Final</version.org.infinispan>
+        <version.org.infinispan>10.1.8.Final</version.org.infinispan>
         <version.org.infinispan.protostream>4.3.2.Final</version.org.infinispan.protostream>
         <version.org.jasypt>1.9.3</version.org.jasypt>
         <version.org.javassist>3.23.2-GA</version.org.javassist>
